core tools common code changes
