+++ src/runtime_src/core/tools/common/XBUtilities.cpp
@@ -56,269 +56,21 @@ static const uint32_t FDT_END = 0x9;
 
 // ------ L O C A L  F U N C T I O N S  A N D  S T R U C T S ------------------
 struct fdt_header {
-  uint32_t magic;
-  uint32_t totalsize;
-  uint32_t off_dt_struct;
-  uint32_t off_dt_strings;
-  uint32_t off_mem_rsvmap;
-  uint32_t version;
-  uint32_t last_comp_version;
-  uint32_t boot_cpuid_phys;
-  uint32_t size_dt_strings;
-  uint32_t size_dt_struct;
+    uint32_t magic;
+    uint32_t totalsize;
+    uint32_t off_dt_struct;
+    uint32_t off_dt_strings;
+    uint32_t off_mem_rsvmap;
+    uint32_t version;
+    uint32_t last_comp_version;
+    uint32_t boot_cpuid_phys;
+    uint32_t size_dt_strings;
+    uint32_t size_dt_struct;
 };
-
-
-// ------ N A M E S P A C E ---------------------------------------------------
-using namespace XBUtilities;
-
-// ------ S T A T I C   V A R I A B L E S -------------------------------------
-static bool m_bVerbose = false;
-static bool m_bTrace = false;
-static bool m_disableEscapeCodes = false;
-static bool m_bShowHidden = false;
-static bool m_bForce = false;
 
 namespace xq = xrt_core::query;
 
 // ------ F U N C T I O N S ---------------------------------------------------
-void
-XBUtilities::setVerbose(bool _bVerbose)
-{
-  bool prevVerbose = m_bVerbose;
-
-  if ((prevVerbose == true) && (_bVerbose == false))
-    verbose("Disabling Verbosity");
-
-  m_bVerbose = _bVerbose;
-
-  if ((prevVerbose == false) && (_bVerbose == true))
-    verbose("Enabling Verbosity");
-}
-
-bool
-XBUtilities::getVerbose()
-{
-  return m_bVerbose;
-}
-
-void
-XBUtilities::setTrace(bool _bTrace)
-{
-  if (_bTrace)
-    trace("Enabling Tracing");
-  else
-    trace("Disabling Tracing");
-
-  m_bTrace = _bTrace;
-}
-
-
-void
-XBUtilities::setShowHidden(bool _bShowHidden)
-{
-  if (_bShowHidden)
-    trace("Hidden commands and options will be shown.");
-  else
-    trace("Hidden commands and options will be hidden");
-
-  m_bShowHidden = _bShowHidden;
-}
-
-bool
-XBUtilities::getShowHidden()
-{
-  return m_bShowHidden;
-}
-
-void
-XBUtilities::setForce(bool _bForce)
-{
-  m_bForce = _bForce;
-
-  if (m_bForce)
-    trace("Enabling force option");
-  else
-    trace("Disabling force option");
-}
-
-bool
-XBUtilities::getForce()
-{
-  return m_bForce;
-}
-
-void
-XBUtilities::disable_escape_codes(bool _disable)
-{
-  m_disableEscapeCodes = _disable;
-}
-
-bool
-XBUtilities::is_escape_codes_disabled() {
-  return m_disableEscapeCodes;
-}
-
-
-void
-XBUtilities::message_(MessageType _eMT, const std::string& _msg, bool _endl, std::ostream & _ostream)
-{
-  static std::map<MessageType, std::string> msgPrefix = {
-    { MT_MESSAGE, "" },
-    { MT_INFO, "Info: " },
-    { MT_WARNING, "Warning: " },
-    { MT_ERROR, "Error: " },
-    { MT_VERBOSE, "Verbose: " },
-    { MT_FATAL, "Fatal: " },
-    { MT_TRACE, "Trace: " },
-    { MT_UNKNOWN, "<type unknown>: " },
-  };
-
-  // A simple DRC check
-  if (_eMT > MT_UNKNOWN) {
-    _eMT = MT_UNKNOWN;
-  }
-
-  // Verbosity is not enabled
-  if ((m_bVerbose == false) && (_eMT == MT_VERBOSE)) {
-      return;
-  }
-
-  // Tracing is not enabled
-  if ((m_bTrace == false) && (_eMT == MT_TRACE)) {
-      return;
-  }
-
-  _ostream << msgPrefix[_eMT] << _msg;
-
-  if (_endl == true) {
-    _ostream << std::endl;
-  }
-}
-
-void
-XBUtilities::message(const std::string& _msg, bool _endl, std::ostream & _ostream)
-{
-  message_(MT_MESSAGE, _msg, _endl, _ostream);
-}
-
-void
-XBUtilities::info(const std::string& _msg, bool _endl)
-{
-  message_(MT_INFO, _msg, _endl);
-}
-
-void
-XBUtilities::warning(const std::string& _msg, bool _endl)
-{
-  message_(MT_WARNING, _msg, _endl);
-}
-
-void
-XBUtilities::error(const std::string& _msg, bool _endl)
-{
-  message_(MT_ERROR, _msg, _endl);
-}
-
-void
-XBUtilities::verbose(const std::string& _msg, bool _endl)
-{
-  message_(MT_VERBOSE, _msg, _endl);
-}
-
-void
-XBUtilities::fatal(const std::string& _msg, bool _endl)
-{
-  message_(MT_FATAL, _msg, _endl);
-}
-
-void
-XBUtilities::trace(const std::string& _msg, bool _endl)
-{
-  message_(MT_TRACE, _msg, _endl);
-}
-
-
-
-void
-XBUtilities::trace_print_tree(const std::string & _name,
-                              const boost::property_tree::ptree & _pt)
-{
-  if (m_bTrace == false) {
-    return;
-  }
-
-  XBUtilities::trace(_name + " (JSON Tree)");
-
-  std::ostringstream buf;
-  boost::property_tree::write_json(buf, _pt, true /*Pretty print*/);
-  XBUtilities::message(buf.str());
-}
-
-
-std::string
-XBUtilities::wrap_paragraphs( const std::string & unformattedString,
-                              unsigned int indentWidth,
-                              unsigned int columnWidth,
-                              bool indentFirstLine) {
-  std::vector<std::string> lines;
-
-  // Process the string
-  std::string workingString;
-
-  for (const auto &entry : unformattedString) {
-    // Do we have a new line added by the user
-    if (entry == '\n') {
-      lines.push_back(workingString);
-      workingString.clear();
-      continue;
-    }
-
-    workingString += entry;
-
-    // Check to see if this string is too long
-    if (workingString.size() >= columnWidth) {
-      // Find the beginning of the previous 'word'
-      auto index = workingString.find_last_of(" ");
-
-      // None found, keep on adding characters till we find a space
-      if (index == std::string::npos)
-        continue;
-
-      // Add the line and populate the next line
-      lines.push_back(workingString.substr(0, index));
-      workingString = workingString.substr(index + 1);
-    }
-  }
-
-  if (!workingString.empty())
-    lines.push_back(workingString);
-
-  // Early exit, nothing here
-  if (lines.size() == 0)
-    return std::string();
-
-  // -- Build the formatted string
-  std::string formattedString;
-
-  // Iterate over the lines building the formatted string
-  const std::string indention(indentWidth, ' ');
-  auto iter = lines.begin();
-  while (iter != lines.end()) {
-    // Add an indention
-    if (iter != lines.begin() || indentFirstLine)
-      formattedString += indention;
-
-    // Add formatted line
-    formattedString += *iter;
-
-    // Don't add a '\n' on the last line
-    if (++iter != lines.end())
-      formattedString += "\n";
-  }
-
-  return formattedString;
-}
 
 boost::property_tree::ptree
 XBUtilities::get_available_devices(bool inUserDomain)
@@ -391,7 +143,7 @@ str_available_devs(bool _inUserDomain)
   //gather available devices for user to pick from
   std::stringstream available_devs;
   available_devs << "\n Available devices:\n";
-  boost::property_tree::ptree available_devices = get_available_devices(_inUserDomain);
+  boost::property_tree::ptree available_devices = XBUtilities::get_available_devices(_inUserDomain);
   for(auto& kd : available_devices) {
     boost::property_tree::ptree& dev = kd.second;
     available_devs << boost::format("  [%s] : %s\n") % dev.get<std::string>("bdf") % dev.get<std::string>("vbnv");
