+++ src/runtime_src/core/tools/common/XBUtilities.h
@@ -17,6 +17,7 @@
 #ifndef __XBUtilities_h_
 #define __XBUtilities_h_
 
+#include "XBUtilitiesCore.h"
 // Include files
 // Please keep these to the bare minimum
 #include "core/common/device.h"
@@ -32,49 +33,7 @@
 #include <boost/algorithm/string.hpp>
 
 namespace XBUtilities {
-  typedef enum {
-    MT_MESSAGE,
-    MT_INFO,
-    MT_WARNING,
-    MT_ERROR,
-    MT_VERBOSE,
-    MT_FATAL,
-    MT_TRACE,
-    MT_UNKNOWN, 
-  } MessageType;
-
-  /**
-   * Enables / Disables verbosity
-   * 
-   * @param _bVerbose true - enable verbosity
-   *                  false - disable verbosity (default)
-   */
-  void setVerbose(bool _bVerbose);
-  bool getVerbose();
-  void setTrace(bool _bVerbose);
-
-  void setShowHidden(bool _bShowHidden);
-  bool getShowHidden();
-
-  void setForce(bool _bForce);
-  bool getForce();
-
-  void disable_escape_codes( bool _disable );
-  bool is_escape_codes_disabled();  
-
-  void message_(MessageType _eMT, const std::string& _msg, bool _endl = true, std::ostream & _ostream = std::cout);
-
-  void message(const std::string& _msg, bool _endl = true, std::ostream & _ostream = std::cout); 
-  void info(const std::string& _msg, bool _endl = true);
-  void warning(const std::string& _msg, bool _endl = true);
-  void error(const std::string& _msg, bool _endl = true);
-  void verbose(const std::string& _msg, bool _endl = true);
-  void fatal(const std::string& _msg, bool _endl = true);
-  void trace(const std::string& _msg, bool _endl = true);
-
-  void trace_print_tree(const std::string & _name, 
-                        const boost::property_tree::ptree & _pt);
-
+ 
   bool can_proceed(bool force=false);
   void can_proceed_or_throw(const std::string& info, const std::string& error);
 
@@ -82,10 +41,7 @@ namespace XBUtilities {
   void print_exception_and_throw_cancel(const xrt_core::error& e);
   void print_exception_and_throw_cancel(const std::runtime_error& e);
 
-  std::string wrap_paragraphs( const std::string & unformattedString,
-                               unsigned int indentWidth,
-                               unsigned int columnWidth,
-                               bool indentFirstLine);
+
   void collect_devices( const std::set<std::string>  &_deviceBDFs,
                         bool _inUserDomain,
                         xrt_core::device_collection &_deviceCollection);
@@ -93,22 +49,6 @@ namespace XBUtilities {
   boost::property_tree::ptree
   get_available_devices(bool inUserDomain);
   
-  template <typename T>
-  std::vector<T> as_vector( boost::property_tree::ptree const& pt, 
-                            boost::property_tree::ptree::key_type const& key) 
-  {
-    std::vector<T> r;
-
-    boost::property_tree::ptree::const_assoc_iterator it = pt.find(key);
-
-    if( it != pt.not_found()) {
-      for (auto& item : pt.get_child(key)) 
-        r.push_back(item.second);
-    }
-    return r;
-  }
-
-
    /**
    * get_axlf_section() - Get section from the file passed in
    *
